Typescript initial pass
