+++ lib/react-native-version.ts
@@ -1,7 +1,8 @@
-const { pathOr, is } = require("ramda")
+import { pathOr, is } from "ramda"
+import { GluegunRunContext } from "gluegun"
 
 // the default React Native version for this boilerplate
-const REACT_NATIVE_VERSION = "0.61.2"
+export const REACT_NATIVE_VERSION = "0.61.2"
 
 // where the version lives under gluegun
 const pathToVersion = ["parameters", "options", "react-native-version"]
@@ -19,12 +20,7 @@ const getVersionFromContext = pathOr(REACT_NATIVE_VERSION, pathToVersion)
  *
  * @param {*} context - The gluegun context.
  */
-const getReactNativeVersion = (context = {}) => {
-  const version = getVersionFromContext(context)
+export const getReactNativeVersion = (context: GluegunRunContext) => {
+  const version = getVersionFromContext(context || {})
   return is(String, version) ? version : REACT_NATIVE_VERSION
-}
-
-module.exports = {
-  REACT_NATIVE_VERSION,
-  getReactNativeVersion,
 }
